refactor: introduce Itest.PlasmaFramework to gather similar tools
