+++ priv/cabbage/apps/itest/lib/account.ex
@@ -3,34 +3,8 @@ defmodule Itest.Account do
     Maintaining used accounts state so that we're able to run tests multiple times.
   """
 
-  alias Itest.Transactions.Currency
   alias Itest.Transactions.Encoding
   import Itest.Poller, only: [wait_on_receipt_confirmed: 1]
-
-  def plasma_framework() do
-    contracts = parse_contracts()
-
-    contracts["CONTRACT_ADDRESS_PLASMA_FRAMEWORK"]
-    |> EIP55.encode()
-    |> elem(1)
-  end
-
-  @ether_vault_id 1
-  def vault_id(currency) do
-    ether = Currency.ether()
-
-    case currency do
-      ^ether -> @ether_vault_id
-    end
-  end
-
-  def vault(currency) do
-    ether = Currency.ether()
-
-    case currency do
-      ^ether -> get_vault(@ether_vault_id)
-    end
-  end
 
   @spec take_accounts(integer()) :: map()
   def take_accounts(number_of_accounts) do
@@ -92,42 +66,4 @@ defmodule Itest.Account do
   end
 
   defp hash(message), do: ExthCrypto.Hash.hash(message, ExthCrypto.Hash.kec())
-
-  defp get_vault(id) do
-    data = ABI.encode("vaults(uint256)", [id])
-    {:ok, result} = Ethereumex.HttpClient.eth_call(%{to: plasma_framework(), data: Encoding.to_hex(data)})
-
-    result
-    |> Encoding.to_binary()
-    |> ABI.TypeDecoder.decode([:address])
-    |> hd()
-    |> Encoding.to_hex()
-    |> EIP55.encode()
-    |> elem(1)
-  end
-
-  # taken from the plasma-contracts deployment snapshot
-  # this parsing occurs in several places around the codebase
-  defp parse_contracts() do
-    local_umbrella_path = Path.join([File.cwd!(), "../../../../", "localchain_contract_addresses.env"])
-
-    contract_addreses_path =
-      case File.exists?(local_umbrella_path) do
-        true ->
-          local_umbrella_path
-
-        _ ->
-          # CI/CD
-          Path.join([File.cwd!(), "localchain_contract_addresses.env"])
-      end
-
-    contract_addreses_path
-    |> File.read!()
-    |> String.split("\n", trim: true)
-    |> List.flatten()
-    |> Enum.reduce(%{}, fn line, acc ->
-      [key, value] = String.split(line, "=")
-      Map.put(acc, key, value)
-    end)
-  end
 end
